Removed unnecessary redundant performMcRun method in McRunnable.

+++ src/edu/lclark/orego/core/Board.java
@@ -341,7 +341,7 @@ public final class Board implements Serializable {
 	/**
 	 * Returns the number of consecutive passes ending the move sequence so far.
 	 */
-	public int getPasses() {
+	public short getPasses() {
 		return passes;
 	}
